@@ -1,4 +1,4 @@
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
@@ -22,23 +22,24 @@ target_link_libraries(BIP151RekeyTest
     ${GTEST_LIB}
     ArmoryCommon
     ${OS_SPECIFIC_LIBS}
+    ${WINSOCK_LIB}
 )
 
-add_executable(ContainerTests
-    ContainerTests.cpp
-    ${TEST_SOURCES}
-)
-set_target_properties(ContainerTests
-    PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}
-)
-target_include_directories(ContainerTests
-    ${LIBARMORYCOMMON_INCLUDE_DIRECTORIES}
-)
-target_link_libraries(ContainerTests
-    ${GTEST_LIB}
-    ArmoryCommon
-    ${OS_SPECIFIC_LIBS}
-)
+#add_executable(ContainerTests
+#    ContainerTests.cpp
+#    ${TEST_SOURCES}
+#)
+#set_target_properties(ContainerTests
+#    PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}
+#)
+#target_include_directories(ContainerTests
+#    ${LIBARMORYCOMMON_INCLUDE_DIRECTORIES}
+#)
+#target_link_libraries(ContainerTests
+#    ${GTEST_LIB}
+#    ArmoryCommon
+#    ${OS_SPECIFIC_LIBS}
+#)
 
 add_executable(CppBlockUtilsTests
     CppBlockUtilsTests.cpp
@@ -54,6 +55,9 @@ target_link_libraries(CppBlockUtilsTests
     ${GTEST_LIB}
     ArmoryCommon
     ${OS_SPECIFIC_LIBS}
+    ${WINSOCK_LIB}
+    capnp
+    kj
 )
 
 # Seems to be outdated - re-enable when fixed
@@ -86,6 +90,9 @@ target_link_libraries(SignerTests
     ${GTEST_LIB}
     ArmoryCommon
     ${OS_SPECIFIC_LIBS}
+    ${WINSOCK_LIB}
+    capnp
+    kj
 )
 
 add_executable(SupernodeTests
@@ -102,6 +109,9 @@ target_link_libraries(SupernodeTests
     ${GTEST_LIB}
     ArmoryCommon
     ${OS_SPECIFIC_LIBS}
+    ${WINSOCK_LIB}
+    capnp
+    kj
 )
 
 add_executable(UtilsTests
@@ -118,11 +128,16 @@ target_link_libraries(UtilsTests
     ${GTEST_LIB}
     ArmoryCommon
     ${OS_SPECIFIC_LIBS}
+    ${WINSOCK_LIB}
+    capnp
+    kj
 )
 
 add_executable(WalletTests
     WalletTests.cpp
     ${TEST_SOURCES}
+    ${LIBBTC_INCLUDE_DIR}/src/trezor-crypto/bip39.c
+    ${LIBBTC_INCLUDE_DIR}/src/trezor-crypto/pbkdf2.c
 )
 set_target_properties(WalletTests
     PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}
@@ -134,6 +149,9 @@ target_link_libraries(WalletTests
     ${GTEST_LIB}
     ArmoryCommon
     ${OS_SPECIFIC_LIBS}
+    ${WINSOCK_LIB}
+    capnp
+    kj
 )
 
 add_executable(ZeroConfTests
@@ -150,4 +168,7 @@ target_link_libraries(ZeroConfTests
     ${GTEST_LIB}
     ArmoryCommon
     ${OS_SPECIFIC_LIBS}
+    ${WINSOCK_LIB}
+    capnp
+    kj
 )
