@@ -154,8 +154,8 @@ class BridgeSocket(object):
       return clientSocket
 
    ####
-   def spawnBridge(self, stringArgs):
-      subprocess.run(["./build/CppBridge", stringArgs])
+   def spawnBridge(self, args: list):
+      subprocess.run(["./build/CppBridge", *args])
 
    #############################################################################
    ## socket write
@@ -495,7 +495,7 @@ class BlockchainService(ProtoWrapper):
 
       fut = self.send(packet)
       reply = fut.getVal()
-      return reply.service.headerData
+      return reply.service.getHeadersByHeight
 
    ####
    def getFeeSchedule(self, strat):
