@@ -359,7 +359,8 @@ void BlockDataViewer::addPublicKey(const SecureBinaryData& pubkey)
 
 ///////////////////////////////////////////////////////////////////////////////
 std::shared_ptr<BlockDataViewer> BlockDataViewer::getNewBDV(
-   const std::string& addr, const std::string& port, const std::string& datadir,
+   const std::string& addr, const std::string& port,
+   const std::filesystem::path& datadir,
    const PassphraseLambda& passLbd, bool ephemeralPeers, bool oneWayAuth,
    std::shared_ptr<RemoteCallback> callbackPtr)
 {
