@@ -371,7 +371,7 @@ namespace AsyncClient
       const std::string& getID(void) const { return bdvID_; }
       static std::shared_ptr<BlockDataViewer> getNewBDV(
          const std::string& addr, const std::string& port,
-         const std::string& datadir, const PassphraseLambda&,
+         const std::filesystem::path& datadir, const PassphraseLambda&,
          bool ephemeralPeers, bool oneWayAuth,
          std::shared_ptr<RemoteCallback> callbackPtr);
 
