@@ -380,49 +380,39 @@ bool DatabaseBuilder::addBlocksToDB(BlockDataLoader& bdl,
    bool fullHints)
 {
    auto blockfilemappointer = bdl.get(fileID);
-   auto ptr = blockfilemappointer->getPtr();
-
-   //ptr is null if we're out of block files
-   if (ptr == nullptr)
+   if (!blockfilemappointer->valid()) {
       return false;
+   }
+   std::map<uint32_t, std::shared_ptr<BlockData>> bdMap;
 
-   map<uint32_t, shared_ptr<BlockData>> bdMap;
-
-   auto getID = [&](const BinaryData&)->uint32_t
+   auto getID = [bcPtr = blockchain_](const BinaryData&)->uint32_t
    {
-      return blockchain_->getNewUniqueID();
+      return bcPtr->getNewUniqueID();
    };
 
-   auto tallyBlocks =
-      [&](const uint8_t* data, size_t size, size_t offset)->bool
+   auto tallyBlocks = [&bdMap, bo, getID, fullHints, fileID](
+      const uint8_t* data, size_t size, size_t offset)->bool
    {
       //deser full block, check merkle
       shared_ptr<BlockData> bd;
       BinaryRefReader brr(data, size);
 
-      try
-      {
+      try {
          bd = BlockData::deserialize(
             data, size, nullptr,
             getID, fullHints ?
             BlockData::CheckHashes::FullHints :
             BlockData::CheckHashes::TxFilters);
-      }
-      catch (const BlockDeserializingException &e)
-      {
+      } catch (const BlockDeserializingException &e) {
          LOGERR << "block deser except: " << e.what();
          LOGERR << "block fileID: " << fileID;
          return false;
-      }
-      catch (const exception &e)
-      {
-         LOGERR << "exception: " << e.what();
+      } catch (const std::exception &e) {
+         LOGERR << "block deser exception: " << e.what();
          return false;
-      }
-      catch (...)
-      {
+      } catch (...) {
          //deser failed, ignore this block
-         LOGERR << "unknown exception";
+         LOGERR << "block deser unknown exception";
          return false;
       }
 
@@ -431,40 +421,37 @@ bool DatabaseBuilder::addBlocksToDB(BlockDataLoader& bdl,
       bd->setOffset(offset);
 
       BlockOffset blockoffset(fileID, offset + bd->size());
-      if (blockoffset > *bo)
+      if (blockoffset > *bo) {
          *bo = blockoffset;
+      }
 
       bdMap.emplace(bd->uniqueID(), bd);
       return true;
    };
 
+   auto ptr = blockfilemappointer->data();
    parseBlockFile(ptr, blockfilemappointer->size(),
       startOffset, tallyBlocks);
 
    //done parsing, add the headers to the blockchain object
    //convert BlockData vector to BlockHeader map first
-   map<HashString, shared_ptr<BlockHeader>> bhmap;
-   for (auto& bd : bdMap)
-   {
+   std::map<HashString, std::shared_ptr<BlockHeader>> bhmap;
+   for (const auto& bd : bdMap) {
       auto bh = bd.second->createBlockHeader();
-      bhmap.insert(move(make_pair(bh->getThisHash(), move(bh))));
+      bhmap.emplace(bh->getThisHash(), std::move(bh));
    }
 
    //add in bulk
    auto insertedBlocks = blockchain_->addBlocksInBulk(bhmap, true);
 
-   if (!fullHints)
-   {
+   if (!fullHints) {
       //process filters
-      if (DBSettings::getDbType() == ARMORY_DB_FULL)
-      {
+      if (DBSettings::getDbType() == ARMORY_DB_FULL) {
          //pull existing file filter bucket from db (if any)
          auto pool = db_->getFilterPoolWriter(fileID);
 
-         if (insertedBlocks.empty())
-         {
-            if (pool.isValid())
-            {
+         if (insertedBlocks.empty()) {
+            if (pool.isValid()) {
                //this block has a filter pool and there is no data to append,
                //we can return
                return true;
@@ -476,9 +463,10 @@ bool DatabaseBuilder::addBlocksToDB(BlockDataLoader& bdl,
          }
 
          //tally all block filters
-         map<uint32_t, shared_ptr<BlockHashVector>> allFilters;
-         for (auto& bdId : insertedBlocks)
+         std::map<uint32_t, std::shared_ptr<BlockHashVector>> allFilters;
+         for (const auto& bdId : insertedBlocks) {
             allFilters.emplace(bdId, bdMap[bdId]->getTxFilter());
+         }
 
          //update bucket
          pool.update(allFilters);
@@ -486,12 +474,11 @@ bool DatabaseBuilder::addBlocksToDB(BlockDataLoader& bdl,
          //update db entry
          db_->putFilterPoolForFileNum(fileID, pool);
       }
-   }
-   else
-   {
+   } else {
       commitAllTxHints(bdMap, insertedBlocks);
-      if (DBSettings::getDbType() == ARMORY_DB_SUPER)
+      if (DBSettings::getDbType() == ARMORY_DB_SUPER) {
          commitAllStxos(bdMap, insertedBlocks);
+      }
    }
 
    return true;
@@ -755,35 +742,31 @@ bool DatabaseBuilder::reparseBlkFiles(unsigned fromID)
 map<BinaryData, shared_ptr<BlockHeader>> DatabaseBuilder::assessBlkFile(
    BlockDataLoader& bdl, unsigned fileID)
 {
-   map<BinaryData, shared_ptr<BlockHeader>> returnMap;
-
-   auto&& blockfilemappointer = bdl.get(fileID);
-   auto ptr = blockfilemappointer->getPtr();
+   std::map<BinaryData, std::shared_ptr<BlockHeader>> returnMap;
+   auto blockfilemappointer = bdl.get(fileID);
 
    //ptr is null if we're out of block files
-   if (ptr == nullptr)
+   if (!blockfilemappointer->valid()) {
       return returnMap;
+   }
+   auto ptr = blockfilemappointer->data();
 
-   vector<shared_ptr<BlockData>> bdVec;
-
+   std::vector<std::shared_ptr<BlockData>> bdVec;
    auto tallyBlocks =
       [&](const uint8_t* data, size_t size, size_t offset)->bool
    {
       //deser full block, check merkle
-      shared_ptr<BlockData> bd;
+      std::shared_ptr<BlockData> bd;
       BinaryRefReader brr(data, size);
 
       auto getID = [this](const BinaryData&)->uint32_t
       { return blockchain_->getNewUniqueID(); };
 
-      try
-      {
+      try {
          bd = BlockData::deserialize(
             data, size, nullptr, getID,
             BlockData::CheckHashes::TxFilters);
-      }
-      catch (...)
-      {
+      } catch (...) {
          //deser failed, ignore this block
          return false;
       }
@@ -793,25 +776,21 @@ map<BinaryData, shared_ptr<BlockHeader>> DatabaseBuilder::assessBlkFile(
 
       //query blockchain object for block by hash
       BlockHeader* bhPtr = nullptr;
-      try
-      {
+      try {
          blockchain_->getHeaderByHash(bd->getHash());
-      }
-      catch (range_error&)
-      {
+      } catch (const range_error&) {
          //catch and continue
       }
 
       //add the block either if we don't have it in our blockchain object,
       //or if the offsets and/or fileID mismatch
-      if (bhPtr != nullptr)
-      {
+      if (bhPtr != nullptr) {
          if (bhPtr->getBlockFileNum() == fileID &&
             bhPtr->getOffset() == offset)
             return true;
       }
 
-      bdVec.emplace_back(move(bd));
+      bdVec.emplace_back(std::move(bd));
       return true;
    };
 
@@ -819,11 +798,10 @@ map<BinaryData, shared_ptr<BlockHeader>> DatabaseBuilder::assessBlkFile(
 
    //done parsing, add the headers to the blockchain object
    //convert BlockData vector to BlockHeader map first
-   map<HashString, shared_ptr<BlockHeader>> bhmap;
-   for (auto& bd : bdVec)
-   {
+   std::map<HashString, std::shared_ptr<BlockHeader>> bhmap;
+   for (const auto& bd : bdVec) {
       auto bh = bd->createBlockHeader();
-      bhmap.insert(make_pair(bh->getThisHash(), bh));
+      bhmap.emplace(bh->getThisHash(), std::move(bh));
    }
 
    return returnMap;
@@ -1116,7 +1094,7 @@ void DatabaseBuilder::verifyTransactions()
                };
 
                auto bdata = BlockData::deserialize(
-                  fileMap->getPtr() + bhPtr->getOffset(),
+                  fileMap->data() + bhPtr->getOffset(),
                   bhPtr->getBlockSize(),
                   bhPtr, getID, BlockData::CheckHashes::NoChecks);
 
@@ -1172,7 +1150,7 @@ void DatabaseBuilder::verifyTransactions()
          };
 
          auto bdata = BlockData::deserialize(
-            fileMap->getPtr() + blockheader->getOffset(),
+            fileMap->data() + blockheader->getOffset(),
             blockheader->getBlockSize(),
             blockheader, getID, BlockData::CheckHashes::NoChecks);
 
@@ -1533,7 +1511,7 @@ void DatabaseBuilder::checkTxHintsIntegrity()
             LOGINFO << "checking txhints for file " << counter;
 
          auto blockfilemappointer = bdl.get(counter);
-         auto ptr = blockfilemappointer->getPtr();
+         auto ptr = blockfilemappointer->data();
          if (ptr == nullptr)
             return;
 
