@@ -5,7 +5,7 @@
 //  See LICENSE-ATI or http://www.gnu.org/licenses/agpl.html                  //
 //                                                                            //
 //                                                                            //
-//  Copyright (C) 2016, goatpig                                               //
+//  Copyright (C) 2016-2024, goatpig                                          //
 //  Distributed under the MIT license                                         //
 //  See LICENSE-MIT or https://opensource.org/licenses/MIT                    //
 //                                                                            //
@@ -15,6 +15,7 @@
 #define _H_DBUTILS
 
 #include <string>
+#include <filesystem>
 #include "BinaryData.h"
 
 enum BLKDATA_TYPE
@@ -43,14 +44,7 @@ enum DB_PREFIX
    DB_PREFIX_TEMPSCRIPT
 };
 
-struct FileMap
-{
-   size_t size_;
-   uint8_t* filePtr_ = nullptr;
-
-   void unmap(void);
-};
-
+////////
 class DBUtils
 {
 public:
@@ -142,15 +136,45 @@ public:
 
    static BinaryData getFilterPoolKey(uint32_t filenum);
    static BinaryData getMissingHashesKey(uint32_t id);
-
-   static bool fileExists(const std::string& path, int mode);
-
-   static FileMap getMmapOfFile(const std::string&, bool write = false);
-   
-   static int removeDirectory(const std::string&);
-   static bool isFile(const std::string& path);
-   static bool isDir(const std::string& path);
-
    static BinaryDataRef getDataRefForPacket(const BinaryDataRef& packet);
+};
+
+namespace FileUtils
+{
+   //used for blk file parsing
+   class FileMap
+   {
+   private:
+      uint8_t* ptr_ = nullptr;
+      size_t size_ = 0;
+
+   public:
+      FileMap(const std::filesystem::path&, bool write=false);
+      ~FileMap(void);
+      const size_t& size(void) const;
+      uint8_t* ptr(void) const;
+      bool isValid(void) const;
+   };
+
+   ////
+   bool fileExists(const std::filesystem::path&, int);
+   bool isFile(const std::filesystem::path&);
+   bool isDir(const std::filesystem::path&);
+   size_t getFileSize(const std::filesystem::path&);
+
+   //core blk file naming pattern
+   std::filesystem::path getBlkFilename(
+      const std::filesystem::path&, uint32_t);
+
+   //used in tests
+   bool copy(const std::filesystem::path&,
+      const std::filesystem::path&,
+      size_t nBytes=SIZE_MAX);
+   bool append(const std::filesystem::path&,
+      const std::filesystem::path&);
+
+   int removeDirectory(const std::filesystem::path&);
+   void createDirectory(const std::filesystem::path&);
+   std::filesystem::path getUserHomePath(void);
 };
 #endif
